requirements.txt
aiohttp==3.12.15
celery==5.5.3
daphne==4.2.1
Django==5.2.6
django-celery-beat==2.8.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
drf-spectacular==0.28.0
psycopg2-binary==2.9.10
